// client/ipc.ts
export enum IPC_EVENTS {
    GET_ORDERS = 'getOrders',
    GET_ORDER = 'getOrder',
    UPDATE_ORDER = 'updateOrder',
    DELETE_ORDER = 'deleteOrder',
    GET_SCULPTURES = 'getSculptures',
    GET_SCULPTURE = 'getSculpture',
    UPDATE_SCULPTURE = 'updateSculpture',
    DELETE_SCULPTURE = 'deleteSculpture'
}